roomsample -> samples.instrumentation
