+++ examples/android-instrumentation-sample/.gitignore
