+++ examples/android-instrumentation-sample/build.gradle.kts
