+++ examples/android-instrumentation-sample/debug.keystore
