+++ examples/android-instrumentation-sample/sentry.properties
