+++ examples/android-instrumentation-sample/src/main/assets/tracks.db
