+++ examples/android-instrumentation-sample/src/main/res/layout/activity_lyrics.xml
