+++ examples/android-instrumentation-sample/src/main/res/layout/activity_main.xml
