+++ examples/android-instrumentation-sample/src/main/res/menu/edit.xml
