+++ examples/android-instrumentation-sample/src/main/res/menu/main.xml
