+++ examples/android-room-lib/src/main/AndroidManifest.xml
@@ -1,2 +1,2 @@
 <?xml version="1.0" encoding="utf-8"?>
-<manifest package="io.sentry.android.roomsample.lib" />
+<manifest package="io.sentry.android.instrumentation.lib" />
