--- examples/android-room/src/main/AndroidManifest.xml
@@ -1,40 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="io.sentry.android.roomsample">
-
-    <application android:name=".SampleApp">
-        <activity
-            android:name=".ui.MainActivity"
-            android:theme="@style/Theme.AppCompat.NoActionBar">
-            <intent-filter>
-                <action android:name="android.intent.action.MAIN" />
-                <category android:name="android.intent.category.LAUNCHER" />
-            </intent-filter>
-        </activity>
-
-        <activity
-            android:name=".ui.EditActivity"
-            android:theme="@style/Theme.AppCompat.NoActionBar" />
-
-        <activity
-            android:name=".ui.LyricsActivity"
-            android:theme="@style/Theme.AppCompat.NoActionBar" />
-
-        <meta-data
-            android:name="io.sentry.dsn"
-            android:value="https://1053864c67cc410aa1ffc9701bd6f93d@o447951.ingest.sentry.io/5428559" />
-
-        <meta-data
-            android:name="io.sentry.debug"
-            android:value="true" />
-
-        <meta-data
-            android:name="io.sentry.traces.sample-rate"
-            android:value="1.0" />
-
-        <meta-data
-            android:name="io.sentry.traces.activity.auto-finish.enable"
-            android:value="false" />
-    </application>
-
-</manifest>
