--- examples/android-room/src/main/java/io/sentry/android/roomsample/ui/LyricsActivity.kt
@@ -1,60 +1,0 @@
-package io.sentry.android.roomsample.ui
-
-import android.annotation.SuppressLint
-import android.os.Bundle
-import android.widget.EditText
-import androidx.activity.ComponentActivity
-import androidx.appcompat.widget.Toolbar
-import io.sentry.Sentry
-import io.sentry.SpanStatus
-import io.sentry.android.roomsample.R
-import io.sentry.android.roomsample.data.Track
-import io.sentry.android.roomsample.util.Filesystem
-import java.io.File
-
-@SuppressLint("SetTextI18n")
-class LyricsActivity : ComponentActivity() {
-    private lateinit var lyricsInput: EditText
-    private lateinit var filesystem: Filesystem
-    private lateinit var track: Track
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_lyrics)
-
-        val transaction = Sentry.startTransaction(
-            "Track Interaction",
-            "ui.action.lyrics",
-            true
-        )
-
-        lyricsInput = findViewById(R.id.lyrics)
-        val toolbar = findViewById<Toolbar>(R.id.toolbar)
-
-        track = intent.getSerializableExtra(TRACK_EXTRA_KEY) as Track
-        filesystem = intent.getSerializableExtra(FILESYSTEM_EXTRA_KEY) as Filesystem
-        toolbar.title = "Lyrics for ${track.name}"
-
-        val dir = File("$filesDir${File.separatorChar}lyrics")
-        dir.mkdirs()
-
-        lyricsInput.setText(filesystem.read(this, "${track.id}.txt"))
-        transaction.finish(SpanStatus.OK)
-    }
-
-    override fun onBackPressed() {
-        val transaction = Sentry.getSpan() ?: Sentry.startTransaction(
-            "Track Interaction",
-            "ui.action.lyrics_finish",
-            true
-        )
-        filesystem.write(this, "${track.id}.txt", lyricsInput.text.toString())
-        transaction.finish(SpanStatus.OK)
-        super.onBackPressed()
-    }
-
-    companion object {
-        const val TRACK_EXTRA_KEY = "LyricsActivity.Track"
-        const val FILESYSTEM_EXTRA_KEY = "LyricsActivity.Filesystem"
-    }
-}
