--- examples/android-room/src/main/res/layout/track_row.xml
@@ -1,57 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<io.sentry.android.roomsample.ui.list.TrackRow xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="wrap_content"
-    android:gravity="center_vertical"
-    android:orientation="horizontal"
-    android:padding="10dp">
-
-    <LinearLayout
-        android:layout_width="0dp"
-        android:layout_height="wrap_content"
-        android:layout_weight="1"
-        android:orientation="vertical">
-
-        <TextView
-            android:id="@+id/track_name"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:textColor="?android:attr/textColorPrimary" />
-
-        <TextView
-            android:id="@+id/band_name"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:textColor="?android:attr/textColorSecondary" />
-    </LinearLayout>
-
-    <TextView
-        android:id="@+id/track_duration"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:textColor="?android:attr/textColorPrimary" />
-
-    <ImageView
-        android:id="@+id/track_info"
-        android:layout_width="24dp"
-        android:layout_height="24dp"
-        android:layout_marginHorizontal="16dp"
-        android:src="@android:drawable/ic_menu_info_details"
-        tools:ignore="ContentDescription" />
-
-    <ImageView
-        android:id="@+id/edit_track"
-        android:layout_width="24dp"
-        android:layout_height="24dp"
-        android:layout_marginEnd="16dp"
-        android:src="@android:drawable/ic_menu_edit"
-        tools:ignore="ContentDescription" />
-
-    <ImageView
-        android:id="@+id/delete_track"
-        android:layout_width="24dp"
-        android:layout_height="24dp"
-        android:src="@android:drawable/ic_menu_delete"
-        tools:ignore="ContentDescription" />
-</io.sentry.android.roomsample.ui.list.TrackRow>
