+++ plugin-build/src/main/kotlin/io/sentry/android/gradle/instrumentation/util/SentryPackageNameUtils.kt
@@ -7,7 +7,6 @@ import com.android.build.api.instrumentation.ClassContext
 fun ClassContext.isSentryClass(): Boolean =
     when {
         currentClassData.className.startsWith("io.sentry") &&
-            !currentClassData.className.startsWith("io.sentry.android.roomsample") &&
             !currentClassData.className.startsWith("io.sentry.samples") &&
             !currentClassData.className.startsWith("io.sentry.mobile") -> true
         else -> false
