+++ settings.gradle.kts
@@ -21,7 +21,7 @@ rootProject.name = ("sentry-android-gradle-plugin-composite-build")
 include(":examples:android-gradle")
 include(":examples:android-gradle-kts")
 include(":examples:android-ndk")
-include(":examples:android-room")
+include(":examples:android-instrumentation-sample")
 include(":examples:android-room-lib")
 includeBuild("plugin-build") {
     dependencySubstitution {
